Refactored seeders to reflect new table structure
