+++ database/seeds/DatabaseSeeder.php
@@ -1,10 +1,17 @@
 <?php
 
-use Illuminate\Database\Seeder;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
 {
+    private $tables = [
+        'roles',
+        'schools',
+        'questions',
+        'question_sets',
+        'question_question_set',
+    ];
     /**
      * Run the database seeds.
      *
@@ -13,13 +20,23 @@ class DatabaseSeeder extends Seeder
     public function run()
     {
         Model::unguard();
+        $this->truncateTables();
 
-        // $this->call(UserTableSeeder::class);
         $this->call(RoleTableSeeder::class);
         $this->call(SchoolTableSeeder::class);
         $this->call(QuestionTableSeeder::class);
-        $this->call(QuestionMetadataTableSeeder::class);
+        $this->call(QuestionSetTableSeeder::class);
+        $this->call(QuestionQuestionSetTableSeeder::class);
 
         Model::reguard();
+    }
+
+    private function truncateTables()
+    {
+        DB::statement('SET FOREIGN_KEY_CHECKS=0');
+        foreach ($this->tables as $table) {
+            DB::statement("truncate $table");
+        }
+        DB::statement('SET FOREIGN_KEY_CHECKS=1');
     }
 }
